feat: Add FirebaseProviderService
Used to deliver data from Firebase Database. This class keeps
ValueEventListeners alive that have any items that are listening to
changes in the underlying data.

It will reduce the number of network calls to Firebase Database and
reduce the data usage for the service, while maintaining fresher data
for UI components.