fix: Fix crash when registering ModelChangeListener
When registering a ModelChangeListener in FirebaseProviderService for a
SmartValueEventListener that already has another listener, it used to
attempt to pass the loaded model to the ModelChangeListener. However,
if the second listener to registered before the data has been returned
by Firebase, then it attempts to pass a null data value, causing a
NPE in onModelReady.

Fixed the issue by adding a null check prior to passing the data to the
ModelChangeListener.