feat: Update TestUserFragment
Fixed the logout function so that after logging out, another user can
properly sign in.

Added ability to handle a new Fragment with no arguments to either load
the logged in user's Firebase profile or go to AccountActivity if the
user is not logged in.

Modified how the ModelChangeListeners are loaded, registered, and un-
registered.

Added real-time notifications when Chats are updated.