fix: Fix crash when unbinding FPS
Even with synchronization, there is a lag between when the Service is
unbound and mBound is marked false. Due to this difference, there was a
chance that unbind() would be called multiple times before the boolean
was flipped, causing an IllegalArgumentException.

Fixed it by immediately setting mBound to false when unbindService() is
called.