fix: Fix Adapter not loading items
The method for getting the item count for the Adapter was still
returning the size of the List and not the new SortedList that had
superseded it.

Also removed the method for checking whether the Adapter is empty
because there is already a method for getting the item count for the
Adapter.