fix: Fix layout for list_item_search_user
When searching for a username, the layout would become taller and
taller due to the constraints on the username TextView being matched
to the constraints of the name TextView. However, because the name
TextView was set to GONE, it caused the username TextView to become
taller to compensate.

Changed the constraints of the username so it was not longer tied to
the name TextView.