fix: Fix crash when loading Authors in ChatFragment
When adding authors to the local database, the function for generating
the ContentValues checked for whether the Author had a corresponding
image. If the Author had a corresponding image, it attempted to add the
Uri of the image to the ContentValues.

However, due to a bug in the way that the hasImage field for Author
works, it flips to positive only when the image is downloaded to the
device. When the Author's profile is updated on Firebase it flips to
true as well. When the profile is re-downloaded, it attempts to get a
Uri for an image that doesn't exist.