fix: Fix algorithm for updating GlideImageSignature
The algorithm for calculating the time since the last update of the
GlideImageSignature only calculated the number of minutes of the hour
since the last update. This meant that it could never be greater than
the three hour trigger to update the GlideImageSignature.

Changed the calculation to calculate the absolute number of minutes
rather than the relative number of minutes of the hour.

(cherry picked from commit 56684f7)