fix: Fix ProgressBar not hiding in UserFragment
When rotating the device, UserFragment re-showed the ProgressBar in its
onConnected() method. However, since there was no need to re-download
the information as it was cached in onSavedInstanceState(), there was
no call to hide the ProgressBar.

Removed the call to show the ProgressBar in onConnected() to fix the
issue.

(cherry picked from commit ed4c2d1)