fix: Fix crash when returning to RequestFragment
When returning to RequestFragment after responding to a friend request,
the Fragment would crash because it attempted to update a position in
the RequestAdapter that did not exist.

Fixed the issue by changing all insertion/removal/changing of items to
trigger notifyDataSetChanged() instead of the individual
notifyDataRange() methods.

(cherry picked from commit 9433827)